# swagger.yaml
openapi: 3.1.0
info:
  title: Resi's Project Web Course API
  description: This is Resis's Project Web Course API docs
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
  - url: https://staging.akuanalis.com/api/v1
  - url: https://dev-api.akuanalis.com/api/v1
  - url: https://akuanalis.com/api/v1
...
paths:
...
  /admin/auth/register:
    post:
      tags:
        - auth
      summary: Register for admin role
      operationId: adminRegister
      requestBody:
        description: Register form
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationForm'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: invalid email/password
  /instructors/auth/register:
    post:
      tags:
        - auth
      summary: Register for instructors role
      operationId: instructorRegister
      requestBody:
        description: Register form
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationForm'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: invalid email/password
  /users/auth/register:
    post:
      tags:
        - auth
      summary: Register for user role
      operationId: userRegister
      requestBody:
        description: Register form
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationForm'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: invalid email/password
...
  /instructors/auth/data:
    post:
      tags:
        - auth
      summary: To complete instructor data
      operationId: completeInstructorData
      security:
        - adminSessionCookie: []
      requestBody:
        description: Instructor data form
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorDataForm'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instructor'
        '400':
          description: invalid data
    put:
      tags:
        - auth
      summary: To update instructor data
      operationId: updateInstructorData
      security:
        - adminSessionCookie: []
        - instructorSessionCookie: []
      requestBody:
        description: Instructor data form
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorDataForm'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instructor'
        '400':
          description: invalid data
...
  /admin/auth/login:
    post:
      tags:
        - auth
      summary: Login for admin role
      operationId: loginAdmin
      requestBody:
        description: Admin role login
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginForm'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
          headers:
            Set-Cookie:
              description: Set-Cookie
              schema:
                type: string
        '400':
          description: invalid email/password
  /instructors/auth/login:
    post:
      tags:
        - auth
      summary: Login for instructor role
      operationId: instructorLogin
      requestBody:
        description: Instructor login form
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginForm'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: invalid email/password
  /users/auth/login:
    post:
      tags:
        - auth
      summary: Login for user role
      operationId: userLogin
      requestBody:
        description: User login form
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginForm'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: invalid email/password
  /auth/logout:
    post:
      tags:
        - auth
      summary: Logout for admin and user roles
      operationId: logout
      security:
        - adminSessionCookie: []
        - userSessionCookie: []
        - instructorSessionCookie: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/password/forgot:
    post:
      tags:
        - auth
      summary: To request reset reset token
      operationId: generateResetPasswordToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  examples:
                    - user@mail.com
  /auth/password/{resetToken}:
    put:
      tags:
        - auth
      summary: Reset Password for admin and user roles
      operationId: resetPassword
      parameters:
        - name: resetToken
          in: path
          description: ID that system generate to verify the reset request
          required: true
          schema:
            type: string
      requestBody:
        description: Form to reset password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetForm'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: invalid password
    get:
      tags:
        - auth
      summary: To verify reset token
      operationId: verifyResetToken
      parameters:
        - name: resetToken
          in: path
          description: ID that system generate to verify the reset request
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The reset token is valid
        '400':
          description: invalid reset token
...
components:
...
  schemas:
    RegistrationForm:
      type: object
      properties:
        email:
          type: string
          examples:
            - test@yopmail.com
        password:
          type: string
          examples:
            - Password@312
        password_confirm:
          type: string
          examples:
            - Password@312
        full_name:
          type: string
          examples:
            - Jhon Doe
    LoginForm:
      type: object
      properties:
        email:
          type: string
          examples:
            - example@email.com
        password:
          type: string
      xml:
        name: login_form
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          examples:
            - example@email.com
        password:
          type: string
        full_name:
          type: string
          examples:
            - Jhon Doe
        role:
          type: string
          description: Person has a few roles
          examples:
            - admin
            - user
          enum:
            - admin
            - user
        photo_profile:
          type: string
          examples:
            - adslfkad.jpg
        created_at:
          type: string
          format: date-time
        activated_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      xml:
        name: person
...
    Instructor:
      type: object
      properties:
        person_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        education:
          type: string
          examples:
            - Bachelor of Computer Science
        experience:
          type: string
          examples:
            - 5 years
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
      xml:
        name: instructor
    PasswordResetForm:
      type: object
      properties:
        new_password:
          type: string
        new_password_confirm:
          type: string
...
    InstructorDataForm:
      type: object
      properties:
        person_id:
          type: string
          format: uuid
          examples:
            - 123e4567-e89b-12d3-a456-426614174000
        education:
          type: string
          examples:
            - Bachelor of Computer Science
        experience:
          type: string
          examples:
            - 5 years
...
    ApiResponse:
      type: object
      properties:
        message:
          type: string
      xml:
        name: '##default'
...
  securitySchemes:
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: session
    userSessionCookie:
      type: apiKey
      in: cookie
      name: session
    instructorSessionCookie:
      type: apiKey
      in: cookie
      name: session